@@ -1,6 +1,8 @@
 {
   "name": "graphmob-com/graphmob-api-php",
   "description": "Graphmob API PHP.",
+  "version": "1.0.0",
+  "homepage": "https://gitlab.com/graphmob-com/graphmob-api-php",
   "type": "library",
   "license": "MIT",
   "authors": [
@@ -10,7 +12,7 @@
     }
   ],
   "require": {
-    "php": ">=5.3.0",
+    "php": ">=5.4.0",
     "tcdent/php-restclient": "0.1.6"
   },
   "autoload": {
