@@ -13,28 +13,33 @@ require __DIR__."/ressources/Verify.php";
 
 class Graphmob {
   public $DEFAULT_REST_HOST = "https://api.graphmob.com";
-  public $DEFAULT_REST_BASE_PATH = "/v1/";
+  public $DEFAULT_REST_BASE_PATH = "/v1";
 
-  public function __construct() {
-    $this->auth = array();
+  private function __construct() {
+    $this->auth = [];
 
-    $this->_rest = new RestClient(array(
+    $this->_rest = new RestClient([
+      "user_agent"   => "graphmob-api-php/1.0.0",
       "base_url"     => $this->DEFAULT_REST_HOST.$this->DEFAULT_REST_BASE_PATH,
-      "headers"      => ["Content-Type" => "application/json"],
-      "content_type" => "application/json"
-    ));
+      "content_type" => "application/json",
+      "format"       => "json",
 
-    $this->_rest->register_decoder("json", function($data) {
-      return json_decode($data, TRUE);
-    });
+      "headers"      => [
+        "Content-Type" => "application/json"
+      ]
+    ]);
 
     $this->enrich = new GraphmobEnrich($this);
     $this->search = new GraphmobSearch($this);
     $this->verify = new GraphmobVerify($this);
   }
 
+  protected function _get($resource, $query) {
+    return $this->parent->_rest->get($resource, $query)->decode_response();
+  }
+
   public function setRestHost($host) {
-   $this->_rest->set_option("base_url", $host);
+    $this->_rest->set_option("base_url", $host);
   }
 
   public function authenticate($identifier, $key) {
