@@ -7,6 +7,18 @@
  * Author: Valerian Saliou <valerian@valeriansaliou.name>
  */
 
-// TODO
+require "../vendor/autoload.php";
+
+$client = new Graphmob();
+
+$client->authenticate("ui_a311da78-6b89-459c-8028-b331efab20d5", "sk_f293d44f-675d-4cb1-9c78-52b8a9af0df2");
+
+$data = $client->verify->formatEmail([
+  "email_domain" => "crisp.chat",
+  "first_name" => "Valerian",
+  "last_name" => "Saliou"
+]);
+
+var_dump($data);
 
 ?>
