@@ -7,6 +7,28 @@
  * Author: Valerian Saliou <valerian@valeriansaliou.name>
  */
 
-// TODO
+class GraphmobSearch {
+  private function __construct($parent) {
+    $this->parent = $parent;
+  }
+
+  public function lookupCompanies($query, $page_number = 1) {
+    return $this->parent->_get(
+      "/search/lookup/companies/".$page_number, $query
+    );
+  }
+
+  public function lookupEmails($query, $page_number = 1) {
+    return $this->parent->_get(
+      "/search/lookup/emails/".$page_number, $query
+    );
+  }
+
+  public function suggestCompanies($query, $page_number = 1) {
+    return $this->parent->_get(
+      "/search/suggest/companies/".$page_number, $query
+    );
+  }
+}
 
 ?>
